Conflicts:
	WordScramble/WordScramble.depend
	WordScramble/WordScramble.layout
	WordScramble/bin/Debug/WordScramble
	WordScramble/utils/Utils.cpp
	WordScramble/utils/Utils.h
	WordScramble/view/WordScrambleWindow.cpp

+++ WordScramble/model/AllPossibleWords.h
@@ -15,7 +15,7 @@ class AllPossibleWords
         map<string, string> possibleWords;
         void determineValidWords(vector<string> dictionary, string randomLetters);
         bool isAValidWord(string word, string randomLetters);
-        int letterCount(string word, char letter);
+        size_t letterCount(string word, char letter);
         bool isAPossibleWord(string selectedWord, string randomLetters);
     public:
         AllPossibleWords(vector<string> dictionary, string randomLetters);
